edit

+++ src/main/java/ee/timur/thesis/dto/CommitteeMemberGradeDTO.java
@@ -2,15 +2,17 @@ package ee.timur.thesis.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CommitteeMemberGradeDTO {
 
     private Long id;
     private Long thesisId;
-    private Integer contentScore;
-    private Integer complexityScore;
-    private Integer appearanceScore;
-    private Integer presentationScore;
+    private BigDecimal contentScore;
+    private BigDecimal complexityScore;
+    private BigDecimal appearanceScore;
+    private BigDecimal presentationScore;
     private String name;
     private String secondName;
     private Boolean visibleToOthers;
