+++ src/main/java/ee/timur/thesis/dto/ReviewerGradeDTO.java
@@ -2,15 +2,17 @@ package ee.timur.thesis.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ReviewerGradeDTO {
 
     private Long thesisId;
-    private Integer contentScore;
+    private BigDecimal contentScore;
     private String contentReasoning;
-    private Integer complexityScore;
+    private BigDecimal complexityScore;
     private String complexityReasoning;
-    private Integer appearanceScore;
+    private BigDecimal appearanceScore;
     private String appearanceReasoning;
     private String evaluationSummary;
     private String questions;
